+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.7.4 - Unreleased
 ### Changed
+- Changed contacts controller actions to work with requests other than JSON requests.
 - Changed webhook method to work with new Mailgun webhook payloads.
 
 ## 1.7.3 - 2019-03-26